Problem: code was working against generators
We define the export_macro in models, for zproto. That generates it
into the header file automatically. In this case it's FILEMQ_EXPORT,
which is defined in filemq_library.h by zproto.

Also the change to using czmq.lib instead of libczmq.lib seems
wrong, as I'd rather we're building CZMQ correctly (to produce
libczmq.lib).

Solution: revert 2144420; define export_macro in client/server
models, and regenerate all code.